Making flaskapp testable
