+++ app.py
@@ -1,38 +1,16 @@
 import os
-import time
+from flask import Flask
 
-from flask import Flask, jsonify, request, current_app
-from plexapihandler import PlexApiHandler
 
-ENV_ANALYZE_MEDIA = os.getenv("ANALYZE_MEDIA", "")
-ENV_REFRESH_MEDIA = os.getenv("REFRESH_MEDIA", "true")
 ENV_LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO")
 
-app = Flask(__name__)
-app.logger.setLevel(ENV_LOG_LEVEL)
-sleep = int(os.getenv("SLEEP_INTERVAL", "0"))
-plex_api = PlexApiHandler(os.getenv("PLEX_URL"), os.getenv("PLEX_TOKEN"))
+def create_app():
+    app = Flask(__name__)
+    app.logger.setLevel(ENV_LOG_LEVEL)
+    # Import and register blueprints
+    from api import main_bp
+    app.register_blueprint(main_bp)
+    return app
 
-
-@app.route("/")
-def ping():
-    return "Ping successful"
-
-@app.route("/triggers/manual", methods=["HEAD"])
-def ok():
-    return "Ok"
-
-@app.route("/triggers/manual", methods=["POST", "GET"])
-def trigger():
-    directory = request.args.get("dir")
-
-    if sleep:
-        time.sleep(sleep)
-    
-    current_app.logger.warning("Starting directory scan of: {}".format(directory))
-
-    if directory:
-        metadata_files = plex_api.find_metadata_from_dirs(directory=directory)
-        files_refreshed = plex_api.refresh_metadata(metadata_files, ENV_ANALYZE_MEDIA, ENV_REFRESH_MEDIA)
-        return jsonify(metadata_entries=files_refreshed)
-    return jsonify(metadata_files=[])
+if __name__ == "__main__":
+    create_app().run()
