+++ tests/app_test.py
@@ -1,12 +1,22 @@
 import unittest
-
-from flask import Flask
-
+from dotenv import load_dotenv
+from app import create_app
 
 class AppTest(unittest.TestCase):
     def setUp(self):
-        self.client = Flask("test").app.test_client()
+        load_dotenv()
+        self.app = create_app()
+        self.client = self.app.test_client()
         self.base_url = "http://127.0.0.1"
+
+    def test_ping(self):
+        response = self.client.head(
+            "/triggers/manual",
+            base_url=self.base_url,
+            content_type="application/json"
+        )
+
+        self.assertEquals(response.status_code, 200)
 
     def test_process_triggers(self):
         response = self.client.post(
